% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_functions.R
\name{arrivaltime}
\alias{arrivaltime}
\title{Arrivaltime}
\usage{
arrivaltime(data)
}
\arguments{
\item{data}{dataframe with two columns}
}
\value{
A vector that contains the arrivaltimes
}
\description{
This function transforms a tibble, a data.frame or a
two-column matrix of event values and waiting times for the
next event to a vector containing the total waiting times
for the respective events.
}
\examples{
dat <- data_generation(20, 0.5, 0.5)
arrivaltime(dat)

}
