% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_redundant_photos.R
\name{delete_redundant_photos}
\alias{delete_redundant_photos}
\title{Delete Redundant Photos in the Upload Folder}
\usage{
delete_redundant_photos(directory_to_drive, photos_to_upload)
}
\arguments{
\item{directory_to_drive}{folder for research drive}

\item{photos_to_upload}{folder for photos from SD cards}
}
\value{

}
\description{
After ingesting and renaming new photos, go ahead and delete photos that are duplicates
(i.e., already in the research drive).
}
\examples{
delete_redundant_photos(directory_to_drive, photos_to_upload)
}
