% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_kobo_photos.R
\name{import_kobo_photos}
\alias{import_kobo_photos}
\title{Import newly named kestrel banding photos to the research drive}
\usage{
import_kobo_photos(kobo_kestrel_data, directory_to_drive)
}
\arguments{
\item{directory_to_drive}{}
}
\value{

}
\description{
Import newly named kestrel banding photos to the research drive
}
\examples{
import_kobo_photos(kobo_kestrel_data = '/Users/Jay/Desktop/kobo', directory_to_drive = 'Z:/HeathLab/American Kestrel projects/Full_Cycle_Phenology/Banding Photos/')
}
