% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixdistr_functions.R
\name{mixdistr}
\alias{mixdistr}
\alias{rmixdistr}
\alias{pmixdistr}
\alias{qmixdistr}
\alias{dmixdistr}
\title{Distribution functions and random number generation}
\usage{
rmixdistr(n, tail, ei, scale = 1)

pmixdistr(q, tail, ei, scale = 1, lower.tail = TRUE)

qmixdistr(p, tail, ei, scale = 1, lower.tail = TRUE, log.p = FALSE)

dmixdistr(x, tail, ei, scale = NULL, log.p = FALSE)
}
\arguments{
\item{n}{number of observations.}

\item{tail}{tail parameter}

\item{ei}{extremal index / the weighting}

\item{scale}{scale parameter, default NULL sets scale = ei ^ (-1 / tail)}

\item{lower.tail}{logical; if TRUE, probabilities are \eqn{P(X\leq x)},
otherwise, \eqn{P(X>x)}.}

\item{p}{vector of probabilities.}

\item{log.p}{logical; if \code{TRUE}, probabilitied \eqn{p} are given as \eqn{\log(p)}}

\item{x, q}{vector of quantiles.}
}
\value{
\code{rmixdistr} generates random variables,
\code{pmixdistr} returns the distribution function,
\code{qmixdistr} returns the quantile function, and
\code{dmixdistr} returns the density.
}
\description{
Probability density, cumulative distribution function, quantile function and
random variate generation for a mixture distribuion with the Dirac measure at
point zero and the Mittag-Leffler distribution as components.
}
\details{
The mixed distribution is a mixture of da Dirac measure
in 0 and a Mittag-Leffler distribution with tail parameter
\eqn{\beta} and a scale parameter \eqn{\sigma}. A Mittag-Leffler \eqn{ML(\beta,\sigma)}
distributed random variable \eqn{Z_{\beta}} with parameter
\eqn{\beta\in(0,1]} and \eqn{\sigma=1} can be defined on the positive real numbers
by its Laplace transform
\deqn{\mathcal{L}_{Z_{\beta}}(s)=\frac{1}{1+s^\beta}_.}
For an arbitrary \eqn{\sigma>0} write \eqn{\sigma Z_\beta}.
The mixed distribution has the c.d.f
\deqn{F_{\beta,\theta,\sigma^*}(x)=(1-\theta)\cdot 1_{[0,\infty)}(x)
+\theta\cdot F^*_{\beta,\theta,\sigma^*}(x),}
where \eqn{F^*_{\beta,\theta,\sigma^*}} is the c.d.f
of the \eqn{ML(\beta,\theta^{-1/\beta}\cdot\sigma^*)} distribution,
where \eqn{\sigma^*} corresponds to the input \code{scale}.
}
\examples{
rmixdistr(n = 10, tail = 0.8, ei = 0.5)
pmixdistr(q = 10, tail = 0.8, ei = 0.5)
qmixdistr(p = 0:10 / 10, tail = 0.8, ei = 0.5)
dmixdistr(x = c(0, 1, 10), tail = 0.8, ei = 0.5)
}
\seealso{
\code{\link[MittagLeffleR]{MittagLeffleR}} for the Mittag-Leffler functions
}
