% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimation_functions.R
\name{optim}
\alias{optim}
\alias{optim_bt}
\alias{optim_bts}
\title{Optim function for known scale parameter}
\usage{
optim_bt(
  WW,
  distance_fct,
  ptail,
  rho,
  t = c(0.25, 0.55, 0.85),
  e = c(0.25, 0.55, 0.85),
  a_tail = 0.1,
  a_ei = 0.1,
  method = "L-BFGS-B"
)

optim_bts(
  WW,
  distance_fct,
  t = c(0.25, 0.55, 0.85),
  e = c(0.25, 0.55, 0.85),
  s = NULL,
  a_tail = 0.1,
  a_ei = 0.1,
  method = "L-BFGS-B",
  ...
)
}
\arguments{
\item{WW}{vector of inter-exceedance times}

\item{distance_fct}{R-function calculating the distance between the asymptotic and empirical cdf}

\item{ptail}{P(JJ > u) = k/n (number of exceedances over number of observations)}

\item{rho}{known parameter}

\item{t}{vector of starting values in \eqn{(a_tail,1]} for the tail parameter}

\item{e}{vector of starting values in \eqn{(a_ei,1]} for the extremal index}

\item{a_tail}{lowest possible estimation for the tail parameter}

\item{a_ei}{lowest possible estimation for the extremal index}

\item{method}{algorithm that is used by \code{stats::optim} to minimize the
distance function}

\item{s}{vector of starting values in \eqn{(0,\inf)} for the extremal index}
}
\value{
A tibble containing the estimation for
the extremal index and the tail parameter. \code{value} is the minimum
found
value of the choosen distance.
}
\description{
An optimization function, that estimates the tail
parameter and the extremal index of
the mixture distribution consisting of a
dirac measure in zero and a Mittag Leffler distribution.
The function is using a grid of multiple initialization values
for the extremal index and the tail parameter.
}
\details{
All combinations of \code{e}, \code{t} and \code{s} are used as starting values
to compute the minimum of \code{distance_fct}. The function is minimized
using \code{stats::optim} and every combination \code{e}, \code{t} and \code{s}.
The parameters corresponding to the lowest value found, are returned
as estimation for \code{ei} and \code{tail}.
}
