% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_photo_timeline.R
\name{plot_photo_timeline}
\alias{plot_photo_timeline}
\title{Plot a Timeline of Renamed Nest Box Photos}
\usage{
plot_photo_timeline(
  directory_to_drive,
  renamed_photo_location,
  ggplot_location
)
}
\arguments{
\item{directory_to_drive}{folder for research drive}

\item{renamed_photo_location}{folder where renamed photos are stored}

\item{ggplot_location}{where ggplot will go}
}
\value{

}
\description{
After all of the photos have been renamed, plot a timeline of all of the photos
taken by nest box ID. Photos will be shown in the plot with a '|' mark.
}
\examples{
plot_photo_timeline(directory_to_drive, renamed_photo_location)
}
