% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_kobo_photos.R
\name{rename_kobo_photos}
\alias{rename_kobo_photos}
\title{Rename kobo kestrel photos by banding data}
\usage{
rename_kobo_photos(kobo_kestrel_data)
}
\arguments{
\item{kobo_kestrel_data}{}
}
\value{

}
\description{
Rename kobo kestrel photos by banding data
}
\examples{
rename_kobo_photos(kobo_kestrel_data = '/Users/Jay/Desktop/kobo')

}
