% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_directories.R
\name{setup_directories}
\alias{setup_directories}
\title{Establish directories for upload, raw, and renamed folders}
\usage{
setup_directories(year_for_photos)
}
\arguments{
\item{year_for_photos}{breeding season year (e.g., 2020)}
}
\value{

}
\description{
Create paths to folders where 1) photos that need to be uploaded are stored,
2) raw photos are stored, and 3) renamed photos are stored.
}
\examples{
setup_directories(year_for_photos)
}
