% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_functions.R
\name{thin}
\alias{thin}
\title{Thinning function}
\usage{
thin(data, k = NULL, u = NULL)
}
\arguments{
\item{data}{data.frame, tibble or matrix with two columns.
The first column contains the event magnitudes (class "numeric") and
in the second column contains teh corresponding waiting times (class "numeric")
where the i-th waiting time is the time between the (i-1)-th
and i-th event}

\item{k}{number of exceedances}

\item{u}{threshold}
}
\value{
A tibble with two columns:
\itemize{
\item the exceedances \code{newJJ} and
\item the corresponding waiting times \code{newWW}
}
}
\description{
This function returns all events that
exceed a specified threshold \code{u} from a tibble, a data.frame or a
two-column matrix \code{data}.
Alternatively, the function can also return the \code{k} largest observations.
}
\examples{
data_generation(n = 10000, stability = 0.9, ei = 0.7, wait_dist = "ML") |>
  thin(k = 100)
data_generation(n = 10000, stability = 0.9, ei = 0.7, wait_dist = "ML") |>
  thin(u = 100)

}
