% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_photos.R
\name{update_photos}
\alias{update_photos}
\title{Ingest New Photos and Rename in the Research Drive}
\usage{
update_photos(
  directory_to_drive,
  year_for_photos,
  photos_to_upload,
  raw_photo_location,
  renamed_photo_location
)
}
\arguments{
\item{directory_to_drive}{folder for the research drive}

\item{year_for_photos}{breeding season year}

\item{photos_to_upload}{folder where SD card photos are added}

\item{raw_photo_location}{folder where new photos will be copied}

\item{renamed_photo_location}{folder where photos are renamed for annotation in ViXeN}
}
\value{

}
\description{
Determine which photos are new, copy them to the raw folder, and then
rename them in the research drive. Code will not run if the proper folder
hierarchy is not in place.
}
\examples{
update_photos(directory_to_drive, year_for_photos, photos_to_upload, raw_photo_location, renamed_photo_location)
}
