% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerating_functions.R
\name{data_generation}
\alias{data_generation}
\title{Data Generating Function}
\usage{
data_generation(n, ei = 1, stability = 1, wait_dist = "stable")
}
\arguments{
\item{n}{number of observations}

\item{ei}{true parameter theta in (0,1] (extremal index)}

\item{stability}{The stability parameter of the distribution of the waiting times.
The stability parameter is in \eqn{\mathbb{R}^+\setminus\lbrace 1\rbrace} if \code{wait_dist}
is "pareto" and in (0,1] else.}

\item{wait_dist}{distribution of waiting times. Should be "stable",
"ML" (Mittag-Leffler Distribution) or "pareto".}
}
\value{
A tibble with

\code{JJ} a stationary MAR-process with extremal index \eqn{\theta} (ei).

\code{WW} independent waiting times between the events.
}
\description{
Function that generates a marked point process (JJ\link{i}, WW\link{i}) \eqn{i = 1,...,n}.
The JJ\link{i} is the i-th mark and WW\link{i} is the waiting time between the i-1- and
i-th event.
}
\details{
If the distribution of the waiting times is chosen as "stable" and the stability parameter
is smaller than 1, then the waiting times are stable distributed.
In this case \code{WW} are generated using the
package \code{stabledist}. In the special case of \code{stability}=1, the stable distribution
simplifies to a dirac measure in 1.

For the case of \code{wait_dist="ML"} a Mittag-Leffler distribution where the
tail parameter and the stability parameter are the same. If the stability parameter
is equal to 1, then the Mittag-Leffler distribution corresponds to an exponential
distribution with
rate 1. For the data generation of the Mittag-Leffler distribution
the package \code{MittagLeffleR} is used.

For \code{wait_dist="pareto"} and a stability parameter
\eqn{\alpha} smaller than 1, a Pareto distribution
with shape parameter \eqn{\alpha} and scale parameter
\eqn{\big(1/\Gamma(1-\alpha)\big)^{1/\alpha}} is used
to generate the waiting times. For \eqn{\alpha} larger than 1 the
scale parameter becomes \eqn{(\alpha-1)/\alpha}.For the
generation of the waiting times the Package \code{ReIns} is used.
}
\examples{
dat <- data_generation(20, 0.5, 0.5)
dat
dat2 <- data_generation(15, 0.2, 0.85, wait_dist = "pareto")
dat2
}
