% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datagenerating_functions.R
\name{data_generation}
\alias{data_generation}
\title{Data Generating Function}
\usage{
data_generation(n, tail = 1, ei = 1, wait_dist = "stable", mag_dist = "MAR")
}
\arguments{
\item{n}{number of observations}

\item{tail}{true tail parameter beta in (0,1]}

\item{ei}{true parameter theta in (0,1] (extremal index)}

\item{wait_dist}{distribution of waiting times}

\item{mag_dist}{distribution of event magnitudes}
}
\value{

}
\description{
Data Generating Function
}
