% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimumdistance_functions.R
\name{distance}
\alias{distance}
\alias{distance_cm}
\alias{distance_cm1}
\alias{and}
\alias{distance_cm2}
\alias{distance_cm_mod1}
\alias{distance_cm_mod2}
\title{Minimum Distance Functions}
\usage{
distance_cm(WW, tail, ei, scale = 1)

distance_cm_mod1(WW, tail, ei, scale = 1)

distance_cm_mod2(WW, tail, ei, scale = 1)
}
\arguments{
\item{WW}{A data.frame or a tibble containing inter exceedance times}

\item{tail}{= \eqn{\beta}; tail parameter}

\item{ei}{= \eqn{\theta}; extremal index}

\item{ptail}{= P(JJ > u) = k / n (number of exceedances / number of observations)}
}
\description{
Functions that calculate the Cramér-von-Mises (or modifications of it) distance
between the empirical distribution funtion
of given waiting times and the Cumulative Distribution function
of a mixture distribution of a dirac measure in zero and a
Mittag-Leffler distribution.
}
\details{
For the cumulative distribution function
\eqn{F^*_{\beta,\theta}} of a \eqn{ML\big(\beta,\theta^{-1/\beta}\big)}-distribution,
the cumulative distribution function of the mixture distribution is given as
\deqn{F_{\beta,\theta}(x)=(1-\theta)\cdot\text{I}_{[0,\infty)}
(x)+\theta \cdot F^*_{\beta,\theta}(x).}
The Cramér von Mises distance between the empirical distribution function
\eqn{F_{k_*}} of \code{WW}
and \eqn{F^*_{\beta,\theta}} is defined as
\deqn{\Delta^{[CM]}(F_{k_*},F_{\beta,\theta})
=\int\limits_0^\infty(F_{k_*}(x)-F_{\beta,\theta}(x))^2
dF_{\beta,\theta}(x)
}
where \eqn{k_*=}\code{length(WW)}.
The function \code{distance_cm_mod1} and \code{distance_cm_mod2}
calculating the modified version of
the Cramér von Mises Distance functions
\deqn{
\Delta^{[CMmod1]}(F_{k_*},F_{\beta,\theta})
=\int\limits_0^\infty(F_{k_*}(x)-F_{\beta,\theta}(x))^2
dF^*_{\beta,\theta}(x)
}
respectively
\deqn{
\Delta^{[CMmod2]}(F_{k_*},F_{\beta,\theta})
=\int\limits_0^\infty(\max\lbrace F_{k_*}(x),1-\theta\rbrace
-F_{\beta,\theta}(x))^2
dF^*_{\beta,\theta}(x)
}.
}
