% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_functions.R
\name{interarrivaltime}
\alias{interarrivaltime}
\title{Interarrivaltime}
\usage{
interarrivaltime(data, type = c("mod1", "mod2"))
}
\arguments{
\item{data}{dataframe with two columns}

\item{type}{"mod1" - WW\link{i} is the waiting time between WW\link{i - 1} and WW\link{i}}
}
\value{

}
\description{
Interarrivaltime
}
