% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_functions.R
\name{interarrivaltime}
\alias{interarrivaltime}
\title{Interarrivaltime}
\usage{
interarrivaltime(data)
}
\arguments{
\item{data}{data frame with two columns}
}
\value{
A vector that contains the inter-exceedance times
}
\description{
Interarrivaltime
}
