% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_functions.R
\name{interarrivaltime}
\alias{interarrivaltime}
\title{Interarrivaltime}
\usage{
interarrivaltime(data)
}
\arguments{
\item{data}{dataframe with two columns}
}
\value{

}
\description{
Interarrivaltime
}
