% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_functions.R
\name{magnitudes}
\alias{magnitudes}
\title{magnitudes}
\usage{
magnitudes(data)
}
\arguments{
\item{data}{dataframe with two columns

This function transforms a tibble, a data.frame or a
two-column matrix of event values and waiting times for the
next event to a vector containing the magnitudes
for the respective events.}
}
\value{
A vector that contains the magnitudes
}
\description{
magnitudes
}
\examples{
dat <- data_generation(20, 0.5, 0.5)
magnitudes(dat)

}
