% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_functions.R
\name{thin}
\alias{thin}
\title{Thinning function}
\usage{
thin(data, k = NULL, u = NULL, type = c("mod1", "mod2"))
}
\arguments{
\item{data}{dataframe with two columns}

\item{k}{number of exceedances}

\item{u}{threshold}

\item{type}{"mod1" - WW\link{i} is the waiting time between WW\link{i - 1} and WW\link{i}}
}
\value{

}
\description{
Thinning function
}
