% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thinning_functions.R
\name{thin}
\alias{thin}
\title{Thinning function}
\usage{
thin(data, k = NULL, u = NULL)
}
\arguments{
\item{data}{data.frame, tibble or matrix with two columns.
In the first column are the event magnitudes (JJ) stored and
in the second column are the corresponding waiting times (WW)
where the i-th waiting time is the time between the (i-1)-th
and i-th event}

\item{k}{number of exceedances}

\item{u}{threshold}
}
\value{
A tibble with two columns:
the event values, that survived the thinning (newJJ) and
the corresponding waiting time (newWW)
}
\description{
This function returns all events that
exceed a specified threshold from a tibble, a data.frame or a
two-column matrix.
Alternatively, the function can also return the k observations
with the largest magnitudes.
}
\examples{
dat <- data_generation(20, 0.5, 0.5)
dat
thin(dat, k = 10)
thin(dat, u = 1)

}
